requirements.txt
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
matplotlib>=3.8.0
pandas>=2.1.0
pytest>=8.0.0
pytest_mock>=3.12.0
lxml>=5.0.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
Flask>=3.0.0
gunicorn>=21.0.0
requests>=2.31.0
numpy>=1.23.3
scikit_learn>=1.1.2